Move measurement code into its own module
